swarms-master-rustified/swarms/utils/pdf_to_text_rustified.rs
### Viability of Conversion to Rust
```rust
// The conversion is partially viable. 
// The main challenges lie in finding equivalent Rust libraries for pypdf and try_except_wrapper.
// However, Rust's standard library and external crates can be used to achieve similar functionality.
```

The given Python code uses the `pypdf` library to extract text from a PDF file and a custom `try_except_wrapper` decorator for error handling. To convert this code to Rust, we need to find equivalent Rust libraries or crates that can perform similar tasks.

### Rust Equivalent

We can use the `pdf` crate to read PDF files and the `std::fs::File` module to handle file operations. For error handling, we can use Rust's built-in `Result` and `Error` types.

```rust
// Import necessary crates and modules
use pdf::{Document, Page};
use std::fs::File;
use std::io::Read;
use std::path::Path;

/// Converts a PDF file to a string of text.
///
/// # Arguments
///
/// * `pdf_path` - The path to the PDF file to be converted.
///
/// # Returns
///
/// A `Result` containing the text extracted from the PDF as a `String` if successful, or an `Error` if an error occurs.
pub fn pdf_to_text(pdf_path: &str) -> Result<String, String> {
    // Open the PDF file
    let file = match File::open(pdf_path) {
        Ok(file) => file,
        Err(_) => return Err(format!("The file at {} was not found.", pdf_path)),
    };

    // Create a new PDF document from the file
    let doc = match Document::load_from_file(file) {
        Ok(doc) => doc,
        Err(_) => return Err("An error occurred while reading the PDF file.".to_string()),
    };

    // Initialize an empty string to store the extracted text
    let mut text = String::new();

    // Iterate through each page and extract text
    for page in doc.pages() {
        if let Some(page) = page {
            if let Some(extracted_text) = page.text() {
                text.push_str(&extracted_text);
                text.push('\n');
            }
        }
    }

    // Return the extracted text
    Ok(text)
}

fn main() {
    // Example usage
    match pdf_to_text("test.pdf") {
        Ok(text) => println!("{}", text),
        Err(e) => println!("{}", e),
    }
}
```

### Limitations and Challenges

1.  **Equivalent Library**: The Rust `pdf` crate may not provide the same level of functionality as the Python `pypdf` library. You may need to use additional crates or implement custom PDF parsing logic to achieve the desired results.
2.  **Error Handling**: Rust's error handling is more explicit and verbose than Python's. You need to manually handle errors using `Result`, `Error`, and `Option` types, which can make the code more complex.
3.  **File Handling**: Rust's file handling is more restrictive than Python's. You need to use the `std::fs::File` module and handle file operations explicitly, which can make the code more verbose.
4.  **Dependency Management**: Rust's dependency management system, Cargo, is more explicit than Python's pip. You need to manually manage dependencies and their versions in your `Cargo.toml` file.

### Compatibility with the Rest of the Project

To ensure compatibility with the rest of the project, you should:

1.  **Use Compatible Crates**: Choose crates that are compatible with your project's dependencies and version requirements.
2.  **Maintain Consistent Error Handling**: Use consistent error handling mechanisms throughout your project to ensure that errors are properly propagated and handled.
3.  **Follow Rust Coding Conventions**: Adhere to Rust coding conventions, such as naming conventions, indentation, and commenting, to ensure that your code is readable and maintainable.
4.  **Test Thoroughly**: Test your Rust code thoroughly to ensure that it works correctly and does not introduce any regressions or compatibility issues.