swarms-master-rustified/tests/utils/test_pdf_to_text_rustified.rs
### Conversion Viability Assessment

The provided Python code is mostly viable for conversion to Rust. The main challenges lie in the following areas:
* The `pypdf` library is used for creating a PDF file, which has no direct Rust equivalent. We can use a library like `pdf` to create PDF files in Rust.
* The `pytest` framework is used for testing, which has a Rust equivalent in `rust-test`. We can use the `#[test]` macro to define tests.
* The `tmpdir` fixture is used to create a temporary directory, which can be replaced with the `tempfile` crate in Rust.
* The `pdf_to_text` function is not defined in the provided code, so we'll assume it's defined elsewhere and has a Rust equivalent.

### Potential Risks and Limitations

* The Rust version of the `pdf_to_text` function may not behave exactly like the Python version, potentially leading to differences in test results.
* The error handling in Rust is more explicit than in Python, which may lead to differences in error behavior.

### Rust Conversion

Here's the Rust equivalent of the provided Python code:
```rust
// This conversion is viable, but it requires equivalent functions for pdf_to_text, 
// which is not provided in the original code snippet.

use std::fs::File;
use std::io::{Read, Write};
use std::path::Path;

use pdf::{PdfWriter, PageObject};
use tempfile::TempDir;
use anyhow::{Result, Context};

// Define a function to create a temporary PDF file
fn create_temp_pdf_file(tmp_dir: &TempDir) -> String {
    let mut pdf_writer = PdfWriter::new();
    let pdf_page = PageObject::new();
    pdf_writer.add_page(pdf_page);
    let pdf_file = tmp_dir.path().join("temp.pdf");
    let mut file = File::create(&pdf_file).unwrap();
    pdf_writer.write(&mut file).unwrap();
    pdf_file.to_str().unwrap().to_string()
}

// Define a test module
#[cfg(test)]
mod tests {
    use super::*;

    // Create a temporary directory for testing
    fn temp_dir() -> TempDir {
        TempDir::new().unwrap()
    }

    // Test that pdf_to_text returns a string for a valid PDF file
    #[test]
    fn test_valid_pdf_to_text() -> Result<()> {
        let tmp_dir = temp_dir();
        let pdf_file = create_temp_pdf_file(&tmp_dir);
        let result = pdf_to_text(&pdf_file); // Assume pdf_to_text is defined elsewhere
        assert!(result.is_ok());
        Ok(())
    }

    // Test that pdf_to_text raises an error for a non-existing file
    #[test]
    fn test_non_existing_file() -> Result<()> {
        let result = pdf_to_text("non_existing_file.pdf"); // Assume pdf_to_text is defined elsewhere
        assert!(result.is_err());
        Ok(())
    }

    // Test that pdf_to_text raises an error for a non-PDF file
    #[test]
    fn test_passing_non_pdf_file() -> Result<()> {
        let tmp_dir = temp_dir();
        let file = tmp_dir.path().join("temp.txt");
        let mut file = File::create(&file).unwrap();
        file.write_all(b"This is a test").unwrap();
        let result = pdf_to_text(file.to_str().unwrap()); // Assume pdf_to_text is defined elsewhere
        assert!(result.is_err());
        Ok(())
    }

    // Test that pdf_to_text raises an error for invalid inputs
    #[test]
    fn test_invalid_pdf_to_text() -> Result<()> {
        let invalid_inputs = vec![None, 123, "{}", "[]"];
        for input in invalid_inputs {
            let result = pdf_to_text(input); // Assume pdf_to_text is defined elsewhere
            assert!(result.is_err());
        }
        Ok(())
    }
}
```

Note that the `pdf_to_text` function is assumed to be defined elsewhere and has a Rust equivalent. The error handling in the Rust version is more explicit than in the Python version, which may lead to differences in error behavior.